Firstly, I created the CartShopActivity class in my Android Studio project. This involved creating a new Java class file and extending it from the AppCompatActivity class.
Next, I added the necessary layout XML file for the CartShopActivity, which included a RecyclerView and a button to proceed to the checkout page. The RecyclerView was used to display the items that the user has added to their cart.

To integrate the CartShopActivity with the DrinkActivity, I used an Intent to pass data between the two activities. Specifically, I passed the data of the selected drink items from DrinkActivity to CartShopActivity.

In the CartShopActivity, I created an ArrayList to store the drink items that the user has added to their cart. Each time the user selects a drink item in the DrinkActivity, the data for that item is added to the ArrayList.

To display the items in the RecyclerView, I created a custom RecyclerView adapter class that inflates a layout for each item in the cart ArrayList. The adapter takes in the cart ArrayList as a parameter and binds the data to the layout for each item.

Finally, to save the data in the RecyclerView, I used SharedPreferences to store the cart ArrayList as a JSON string. This allowed me to easily retrieve the cart data when the user returns to the CartShopActivity.

Overall, this process involved creating the necessary Java classes and layout XML files, using Intents to pass data between activities, creating a custom RecyclerView adapter, and using SharedPreferences to save and retrieve data in the CartShopActivity.